feat(rpg-11): replace attribute formulas with additive baseline-60 standard
Every derived attribute in PlayableCharacter now follows attribute =
baseline + sum(weight * (input - neutral)), replacing the previous
multiplicative/power-law model (including the same-day rpg-10 Speed
redesign). All Body input scales move from double/float to discrete
int ranges. CardiovascularCapacity and FatigueRate are removed;
FatigueResistance (inverted semantics) and StaminaRecovery are added,
plus a new Load Capacity group (Light/Heavy/Max/Drag) derived from
Strength. Adds a project skill documenting the standard for future
formula work.

+++ src/main/java/com/keynor/rpg/application/dto/BodyCompositionInput.java
@@ -2,8 +2,8 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.BodyComposition;
 
-public record BodyCompositionInput(double bodyFat, double muscleMass, double dominantFiberType,
-                                    double muscleDistribution, double flexibility) {
+public record BodyCompositionInput(int bodyFat, int muscleMass, int dominantFiberType,
+                                    int muscleDistribution, int flexibility) {
 
     public BodyComposition toDomain() {
         return new BodyComposition(bodyFat, muscleMass, dominantFiberType, muscleDistribution, flexibility);
