+++ src/main/java/com/keynor/rpg/application/dto/BodyCompositionResponse.java
@@ -2,8 +2,8 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.BodyComposition;
 
-public record BodyCompositionResponse(double bodyFat, double muscleMass, double dominantFiberType,
-                                       double muscleDistribution, double flexibility) {
+public record BodyCompositionResponse(int bodyFat, int muscleMass, int dominantFiberType,
+                                       int muscleDistribution, int flexibility) {
 
     public static BodyCompositionResponse from(BodyComposition composition) {
         return new BodyCompositionResponse(composition.getBodyFat(), composition.getMuscleMass(),
