+++ src/main/java/com/keynor/rpg/application/dto/CardiacSystemInput.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.CardiacSystem;
 
-public record CardiacSystemInput(double cardiacOutput) {
+public record CardiacSystemInput(int cardiacOutput) {
 
     public CardiacSystem toDomain() {
         return new CardiacSystem(cardiacOutput);
