+++ src/main/java/com/keynor/rpg/application/dto/CardiacSystemResponse.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.CardiacSystem;
 
-public record CardiacSystemResponse(double cardiacOutput) {
+public record CardiacSystemResponse(int cardiacOutput) {
 
     public static CardiacSystemResponse from(CardiacSystem cardiacSystem) {
         return new CardiacSystemResponse(cardiacSystem.getCardiacOutput());
