+++ src/main/java/com/keynor/rpg/application/dto/GeneticsResponse.java
@@ -2,8 +2,8 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.Genetics;
 
-public record GeneticsResponse(double endomorphy, double mesomorphy, double ectomorphy, double height,
-                                double limbRatio, double boneDensity) {
+public record GeneticsResponse(int endomorphy, int mesomorphy, int ectomorphy, int height,
+                                int limbRatio, int boneDensity) {
 
     public static GeneticsResponse from(Genetics genetics) {
         return new GeneticsResponse(genetics.getEndomorphy(), genetics.getMesomorphy(), genetics.getEctomorphy(),
