+++ src/main/java/com/keynor/rpg/application/dto/NervousSystemResponse.java
@@ -2,7 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.NervousSystem;
 
-public record NervousSystemResponse(double neuralDrive, double neuromuscularEfficiency) {
+public record NervousSystemResponse(int neuralDrive, int neuromuscularEfficiency) {
 
     public static NervousSystemResponse from(NervousSystem nervousSystem) {
         return new NervousSystemResponse(nervousSystem.getNeuralDrive(),
